+++ src/app/data/components/data-browser/data-browser.component.html
@@ -843,8 +843,7 @@
                     <kern-record-card [attributesSortOrder]="attributesSortOrder" [record]="record" [index]="i"
                         [attributes]="attributes" [extendedRecords]="extendedRecords"
                         [similarSearchHelper]="similarSearchHelper" [recordComments]="recordComments"
-                        [dataBrowserModals]="dataBrowserModals" [activeSearchParams]="activeSearchParams"
-                        (preliminaryRecordIds)="storePreliminaryRecordIds($event)">
+                        [dataBrowserModals]="dataBrowserModals" [activeSearchParams]="activeSearchParams">
                     </kern-record-card>
                 </div>
             </div>
